Clean up error responses
There were some areas that were not covered when handling error cases. This extracts out a module to break that functionality a bit and tries to use a few more similar cases.
